[ADZ-2631] Rename Cache to DiskCache
