+++ site/components/src/test/java/uk/nhs/digital/common/components/apispecification/ApiSpecificationRendererDirectiveTest.java
@@ -25,7 +25,6 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import uk.nhs.digital.cache.Cache;
 import uk.nhs.digital.cache.HeavyContentCache;
 import uk.nhs.digital.test.TestLoggerRule;
 import uk.nhs.digital.test.mockito.MockitoSessionTestBase;
